updated init

+++ init.sh
@@ -8,13 +8,21 @@ echo "Initializing and updating git submodules..."
 git submodule update --init
 
 # Download essentia models manually to avoid LFS
-if [ ! -d "essentia/test/models" ]; then
-    echo "Downloading essentia models..."
-    mkdir -p essentia/test/models
-    curl -L https://github.com/MTG/essentia-models/archive/refs/heads/master.zip -o /tmp/models.zip
-    unzip /tmp/models.zip -d /tmp
-    mv /tmp/essentia-models-master/* essentia/test/models/
-    rm -rf /tmp/essentia-models-master /tmp/models.zip
+echo "Do you want to download the ML models (essentia models)? (y/n)"
+read -r answer
+if [ "$answer" = "y" ] || [ "$answer" = "Y" ]; then
+    if [ ! -d "essentia/test/models" ]; then
+        echo "Downloading essentia models..."
+        mkdir -p essentia/test/models
+        curl -L https://github.com/MTG/essentia-models/archive/refs/heads/master.zip -o /tmp/models.zip
+        unzip /tmp/models.zip -d /tmp
+        mv /tmp/essentia-models-master/* essentia/test/models/
+        rm -rf /tmp/essentia-models-master /tmp/models.zip
+    else
+        echo "Essentia models already exist."
+    fi
+else
+    echo "Skipping ML model download."
 fi
 
 if [ $? -eq 0 ]; then
